refactor(command-palette): split CommandPalette into focused modules
Decompose the 847-LOC CommandPalette.tsx into 3 sibling files plus a
slimmed component file. The render body is now 249 LOC focused on
presentation; both behavior maps live in dedicated hooks.

Extracted modules (all under 470 LOC):

- commandPaletteActionHandlers.ts (460) — `useActionHandlers` hook +
  `ActionHandler` type + `dispatchWindowEvent` helper. Builds the
  full Record<commandId, () => void | null> by reading every relevant
  store slice via useShallow + composing per-section maps (navigation,
  edit, selection, layers, view, preview, bins, tools, exportActions).
- commandPaletteContextBoosts.ts (101) — `useContextBoosts` hook
  returning per-command frecency multipliers from current app state
  (selection, layers, preview visibility, paint mode, staging, etc.).
- commandPaletteParts.tsx (71) — `CategoryIcon` (with the
  CATEGORY_ICON_MAP) + `CommandItem` row component.

CommandPalette.tsx: 847 LOC -> 249 LOC. The component file now only
contains the render body and the small wiring (memos for `commands`,
`groupedCommands`, `recentCommands`, `recentCommandIds`,
`selectedCommand`; `handleSelect` callback; the global Escape effect).

4 command-palette test files / 52 tests pass (1.6s targeted run).
